+++ cmake/info.cmake
@@ -63,7 +63,6 @@ message(STATUS " ")
 message(STATUS "MANIFOLD_VERSION:              ${MANIFOLD_VERSION}")
 message(STATUS "MANIFOLD_PAR:                  ${MANIFOLD_PAR}")
 message(STATUS "MANIFOLD_CROSS_SECTION:        ${MANIFOLD_CROSS_SECTION}")
-message(STATUS "MANIFOLD_CROSS_SECTION_BACKEND: ${MANIFOLD_CROSS_SECTION_BACKEND}")
 message(STATUS "MANIFOLD_TEST:                 ${MANIFOLD_TEST}")
 message(STATUS "MANIFOLD_DEBUG:                ${MANIFOLD_DEBUG}")
 message(STATUS "MANIFOLD_ASSERT:               ${MANIFOLD_ASSERT}")
@@ -77,10 +76,6 @@ message(STATUS "MANIFOLD_COMPILE_OPTIONS:      ${MANIFOLD_COMPILE_OPTIONS}")
 message(STATUS "MANIFOLD_LINK_OPTIONS:         ${MANIFOLD_LINK_OPTIONS}")
 message(STATUS "MANIFOLD_DOWNLOADS:            ${MANIFOLD_DOWNLOADS}")
 message(STATUS "MANIFOLD_USE_BUILTIN_TBB:      ${MANIFOLD_USE_BUILTIN_TBB}")
-message(
-  STATUS
-  "MANIFOLD_USE_BUILTIN_CLIPPER2: ${MANIFOLD_USE_BUILTIN_CLIPPER2}"
-)
 message(
   STATUS
   "MANIFOLD_USE_BUILTIN_NANOBIND: ${MANIFOLD_USE_BUILTIN_NANOBIND}"
