+++ cmake/manifold.pc.in
@@ -7,6 +7,6 @@ Name: manifold@PCFILE_LIB_SUFFIX@
 Description: Geometry library for topological robustness
 Version: @MANIFOLD_VERSION@
 URL: https://github.com/elalish/manifold
-Requires-private: @TEMPLATE_OPTIONAL_TBB@ @TEMPLATE_OPTIONAL_CLIPPER@
+Requires-private: @TEMPLATE_OPTIONAL_TBB@
 Libs: -L${libdir} -lmanifold@PCFILE_LIB_SUFFIX@
 Cflags: -I${includedir} -std=c++17
