+++ cmake/manifoldConfig.cmake.in
@@ -15,18 +15,6 @@ set(MANIFOLD_FOUND TRUE)
 
 set(MANIFOLD_SHARED_LIB "@BUILD_SHARED_LIBS@")
 set(MANIFOLD_CROSS_SECTION "@MANIFOLD_CROSS_SECTION@")
-set(MANIFOLD_CROSS_SECTION_BACKEND "@MANIFOLD_CROSS_SECTION_BACKEND@")
-set(MANIFOLD_USE_BUILTIN_CLIPPER2 "@MANIFOLD_USE_BUILTIN_CLIPPER2@")
-# if built as shared lib and using builtin, no need to find clipper2
-# otherwise, we need to find it
-if(
-  MANIFOLD_CROSS_SECTION
-  AND MANIFOLD_CROSS_SECTION_BACKEND STREQUAL "clipper2"
-  AND NOT (MANIFOLD_SHARED_LIB AND MANIFOLD_USE_BUILTIN_CLIPPER2)
-)
-  set(Clipper2_ROOT "${_FIND_ROOT}")
-  find_package(Clipper2 REQUIRED)
-endif()
 set(MANIFOLD_PAR "@MANIFOLD_PAR@")
 set(MANIFOLD_USE_BUILTIN_TBB "@MANIFOLD_USE_BUILTIN_TBB@")
 if(
