+++ cmake/manifoldDeps.cmake
@@ -85,91 +85,6 @@ if(MANIFOLD_PAR)
   endif()
 endif()
 
-# The default CrossSection backend uses Clipper2. The experimental boolean2
-# backend is in-tree and does not need this dependency.
-if(MANIFOLD_CROSS_SECTION AND MANIFOLD_CROSS_SECTION_BACKEND STREQUAL "clipper2")
-  if(NOT MANIFOLD_USE_BUILTIN_CLIPPER2 AND NOT Clipper2_FOUND)
-    find_package(Clipper2 QUIET)
-    if(NOT Clipper2_FOUND AND PKG_CONFIG_FOUND)
-      pkg_check_modules(Clipper2 Clipper2)
-      if(Clipper2_FOUND)
-        add_library(Clipper2 SHARED IMPORTED)
-        set_property(
-          TARGET Clipper2
-          PROPERTY IMPORTED_LOCATION ${Clipper2_LINK_LIBRARIES}
-        )
-        if(WIN32)
-          set_property(
-            TARGET Clipper2
-            PROPERTY IMPORTED_IMPLIB ${Clipper2_LINK_LIBRARIES}
-          )
-        endif()
-        target_include_directories(Clipper2 INTERFACE ${Clipper2_INCLUDE_DIRS})
-      endif(Clipper2_FOUND)
-    endif(NOT Clipper2_FOUND AND PKG_CONFIG_FOUND)
-  endif(NOT MANIFOLD_USE_BUILTIN_CLIPPER2 AND NOT Clipper2_FOUND)
-  if(NOT Clipper2_FOUND)
-    logmissingdep("Clipper2" , "cross_section")
-    set(MANIFOLD_USE_BUILTIN_CLIPPER2 ON)
-    set(CLIPPER2_UTILS OFF)
-    set(CLIPPER2_EXAMPLES OFF)
-    set(CLIPPER2_TESTS OFF)
-    set(
-      CLIPPER2_USINGZ
-      OFF
-      CACHE BOOL
-      "Preempt cache default of USINGZ (we only use 2d)"
-    )
-    # When manifold is built with MANIFOLD_NO_IOSTREAM, also strip
-    # iostream from the bundled Clipper2 — manifold doesn't call any
-    # of Clipper2's stream operators internally, so passing this
-    # through is safe regardless. The CLIPPER2_NO_IOSTREAM macro is
-    # added by the carry-patch below; once Clipper2#1094 lands and
-    # the SHA pin moves past it, the patch drops and the option is
-    # honored natively.
-    if(MANIFOLD_NO_IOSTREAM)
-      set(
-        CLIPPER2_NO_IOSTREAM
-        ON
-        CACHE BOOL
-        "Strip iostream-using overloads from Clipper2 (set by manifold when MANIFOLD_NO_IOSTREAM=ON)"
-        FORCE
-      )
-    endif()
-    FetchContent_Declare(
-      Clipper2
-      GIT_REPOSITORY https://github.com/AngusJohnson/Clipper2.git
-      # Mar 05, 2026
-      GIT_TAG 46f639177fe418f9689e8ddb74f08a870c71f5b4
-      GIT_PROGRESS TRUE
-      SOURCE_SUBDIR
-      CPP
-      # Disable Windows autocrlf on the clone so the carry-patch (which
-      # is LF-only) applies cleanly. Default core.autocrlf=true on
-      # Windows would convert all LFs to CRLFs in the working tree, and
-      # `git apply` then fails on the line-ending mismatch.
-      GIT_CONFIG
-      core.autocrlf=false
-      # Carry-patch: tracks AngusJohnson/Clipper2#1094 (CLIPPER2_NO_IOSTREAM
-      # macro guards). Drops once that PR lands and the SHA pin moves past
-      # it. Applied via wrapper script so re-configures (which re-trigger
-      # PATCH_COMMAND) don't fail when the patch is already applied.
-      PATCH_COMMAND ${CMAKE_COMMAND}
-        -DPATCH_FILE=${CMAKE_CURRENT_LIST_DIR}/patches/0001-clipper2-no-iostream.patch
-        -DSOURCE_DIR=<SOURCE_DIR>
-        -P ${CMAKE_CURRENT_LIST_DIR}/patches/apply-clipper2-patch.cmake
-    )
-    FetchContent_MakeAvailable(Clipper2)
-    set_property(
-      DIRECTORY ${clipper2_SOURCE_DIR}/CPP
-      PROPERTY EXCLUDE_FROM_ALL ${BUILD_SHARED_LIBS}
-    )
-  endif()
-  if(NOT TARGET Clipper2::Clipper2)
-    add_library(Clipper2::Clipper2 ALIAS Clipper2)
-  endif()
-endif()
-
 if(TRACY_ENABLE)
   logmissingdep("tracy" , "TRACY_ENABLE")
   FetchContent_Declare(
