+++ flake.nix
@@ -1,15 +1,10 @@
 {
   inputs.flake-utils.url = "github:numtide/flake-utils";
   inputs.nixpkgs.url = "nixpkgs/nixos-unstable";
-  inputs.clipper2-src = {
-    url = "github:AngusJohnson/Clipper2";
-    flake = false;
-  };
   outputs =
     { self
     , nixpkgs
     , flake-utils
-    , clipper2-src
     }:
     flake-utils.lib.eachDefaultSystem
       (system:
@@ -17,15 +12,6 @@
         manifold-version = "3.5.1";
         pkgs = import nixpkgs {
           inherit system;
-          overlays = [
-            (final: prev: {
-              clipper2 = prev.clipper2.overrideAttrs (_: rec {
-                version = clipper2-src.rev;
-                src = clipper2-src;
-                sourceRoot = "source/CPP";
-              });
-            })
-          ];
         };
         manifold =
           { parallel ? true }: pkgs.stdenv.mkDerivation {
@@ -39,9 +25,7 @@
                 (ps: with ps; [ nanobind trimesh pytest ]))
               gtest
             ]) ++ (if parallel then [ pkgs.onetbb ] else [ ]);
-            buildInputs = with pkgs; [
-              clipper2
-            ];
+            buildInputs = [ ];
             cmakeFlags = [
               "-DMANIFOLD_STRICT=ON"
               "-DMANIFOLD_CBIND=ON"
@@ -62,8 +46,6 @@
           nativeBuildInputs = (with pkgs; [ cmake python3 ]);
           buildInputs = [ pkgs.nodejs ];
           configurePhase = ''
-            cp -r ${clipper2-src} clipper2
-            chmod -R +w clipper2
             mkdir -p .emscriptencache
             export EM_CACHE=$(pwd)/.emscriptencache
             mkdir build
@@ -73,8 +55,7 @@
             -DMANIFOLD_PAR=${if parallel then "ON" else "OFF"} \
             -DMANIFOLD_USE_BUILTIN_TBB=${if parallel then "ON" else "OFF"} \
             -DFETCHCONTENT_SOURCE_DIR_GOOGLETEST=${pkgs.gtest.src} \
-            -DFETCHCONTENT_SOURCE_DIR_TBB=${pkgs.onetbb.src} \
-            -DFETCHCONTENT_SOURCE_DIR_CLIPPER2=../clipper2 ..
+            -DFETCHCONTENT_SOURCE_DIR_TBB=${pkgs.onetbb.src} ..
           '';
           buildPhase = ''
             emmake make -j''${NIX_BUILD_CORES}
@@ -107,7 +88,7 @@
             version = manifold-version;
             src = self;
             propagatedBuildInputs = [ numpy ];
-            buildInputs = with pkgs; [ clipper2 onetbb ];
+            buildInputs = with pkgs; [ onetbb ];
             nativeBuildInputs = with pkgs; [
               cmake
               ninja
@@ -147,7 +128,6 @@
             onetbb
             gtest
             assimp
-            clipper2
             pkg-config
 
             # useful tools
