+++ scripts/benchmark_pr_perf_guard.sh
@@ -1,14 +1,13 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-if [ "$#" -ne 2 ]; then
-  echo "Usage: $0 <clipper2-dir> <repeats>"
+if [ "$#" -ne 1 ]; then
+  echo "Usage: $0 <repeats>"
   exit 2
 fi
 
-CLIPPER2_DIR="$1"
-REPEATS="$2"
+REPEATS="$1"
 
 for variant in base head; do
-  bash ./scripts/run_pr_perf_guard.sh "./wt-${variant}" "$CLIPPER2_DIR" "./bench/${variant}" "$REPEATS"
+  bash ./scripts/run_pr_perf_guard.sh "./wt-${variant}" "./bench/${variant}" "$REPEATS"
 done
