+++ scripts/run_pr_perf_guard.sh
@@ -1,15 +1,14 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-if [ "$#" -ne 4 ]; then
-  echo "Usage: $0 <source-dir> <clipper2-dir> <output-dir> <repeats>"
+if [ "$#" -ne 3 ]; then
+  echo "Usage: $0 <source-dir> <output-dir> <repeats>"
   exit 2
 fi
 
 SRC_DIR="$1"
-CLIPPER2_DIR="$2"
-OUT_DIR="$3"
-REPEATS="$4"
+OUT_DIR="$2"
+REPEATS="$3"
 BUILD_DIR="${OUT_DIR}/build"
 
 cmake \
@@ -20,8 +19,7 @@ cmake \
   -DMANIFOLD_PYBIND=OFF \
   -DMANIFOLD_TEST=ON \
   -DMANIFOLD_PAR=ON \
-  -DMANIFOLD_CROSS_SECTION=OFF \
-  -DFETCHCONTENT_SOURCE_DIR_CLIPPER2="$CLIPPER2_DIR"
+  -DMANIFOLD_CROSS_SECTION=OFF
 
 cmake --build "$BUILD_DIR" --target perfTest
