+++ scripts/test-cmake-subdir.sh
@@ -21,7 +21,7 @@ EOT
 cp -r ../manifold ./
 mkdir build
 cd build
-cmake -DFETCHCONTENT_SOURCE_DIR_CLIPPER2=$CLIPPER2 -DFETCHCONTENT_SOURCE_DIR_TBB=$TBB ..
+cmake -DFETCHCONTENT_SOURCE_DIR_TBB=$TBB ..
 make
 ./testing
