+++ src/svd.h
@@ -301,4 +301,24 @@ inline double SpectralNorm(mat3 A) {
   SVDSet usv = SVD(A);
   return usv.S[0][0];
 }
+
+/**
+ * Returns the largest singular value (spectral norm) of a 2x2 matrix, via the
+ * closed-form eigenvalues of A^T A. The mat3 overload would floor the norm if a
+ * 2D linear map were embedded in 3D.
+ *
+ * @param A The matrix to measure.
+ */
+inline double SpectralNorm(mat2 A) {
+  const vec2 c0 = A[0];
+  const vec2 c1 = A[1];
+  const double m00 = la::dot(c0, c0);
+  const double m11 = la::dot(c1, c1);
+  const double m01 = la::dot(c0, c1);
+  const double trace = m00 + m11;
+  const double det = m00 * m11 - m01 * m01;
+  const double disc2 = trace * trace - 4.0 * det;
+  const double disc = std::sqrt(disc2 > 0.0 ? disc2 : 0.0);
+  return std::sqrt(0.5 * (trace + disc));
+}
 }  // namespace manifold
